src/requirements.txt
boto3
PyJWT
requests
python-jose
